requirements.txt
youtube-transcript-api==0.6.1
google-generativeai>=0.3.0
flask>=2.3.0
python-dotenv>=1.0.0
requests>=2.31.0